# Refs #28459 -- Improved performance of loading DecimalField on SQLite.