--- a/django/db/models/fields/__init__.py
+++ b/django/db/models/fields/__init__.py
@@ -1560,20 +1560,6 @@ class DecimalField(Field):
                 params={'value': value},
             )
 
-    def format_number(self, value):
-        """
-        Format a number into a string with the requisite number of digits and
-        decimal places.
-        """
-        # Method moved to django.db.backends.utils.
-        #
-        # It is preserved because it is used by the oracle backend
-        # (django.db.backends.oracle.query), and also for
-        # backwards-compatibility with any external code which may have used
-        # this method.
-        from django.db.backends import utils
-        return utils.format_number(value, self.max_digits, self.decimal_places)
-
     def get_db_prep_save(self, value, connection):
         return connection.ops.adapt_decimalfield_value(self.to_python(value), self.max_digits, self.decimal_places)